Move GetValidFileName() to the FileUtility class
